Now using SDL for windowing to help with xplatformness

+++ timer.h
@@ -1,12 +1,14 @@
 #pragma once
 
+#include "SDL.h"
+
 namespace timer
 {
     bool init();
 
-    unsigned long long read();
+    Uint64 read();
 
-    unsigned long long to_usec(unsigned long long x);
-    unsigned long long to_msec(unsigned long long x);
-    double to_sec(unsigned long long x);
+    Uint64 to_usec(Uint64 x);
+    Uint64 to_msec(Uint64 x);
+    double to_sec(Uint64 x);
 }
